can `npm test` worker socket
